requirements.txt
# Web框架
fastapi==0.109.0
uvicorn==0.27.0
pydantic>=2.11.0
pydantic-settings>=2.10.0

# 数据库
sqlalchemy>=2.0.30
pymysql==1.1.0
alembic>=1.16.0

# HTTP和异步请求
aiohttp==3.9.1

# 安全防护
slowapi==0.1.9
captcha==0.5.0
pillow==10.4.0

# 基础工具
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.8.0

# 调度器
apscheduler==3.10.4
